backend/pyproject.toml
[tool.poetry]
name = "llm-chat-backend"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
djangorestframework = "^3.15.2"
pydantic = "^2.10.4"
openai = "^1.59.3"
python-dotenv = "^1.0.1"
anthropic = "^0.42.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
